requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlmodel
python-jose[cryptography]
passlib
python-multipart